+++ smawg/tests/test_main.py
@@ -11,7 +11,7 @@ from smawg.basic_rules import (
     Abandon, Conquer, ConquerWithDice, Decline, Deploy, EndTurn, SelectCombo,
     StartRedeployment
 )
-from smawg.cli import _Client, _MaybeDry, _parse_command
+from smawg.cli import _HumanClient, _MaybeDry, _parse_command
 from smawg.default_rules import Action
 from smawg.tests.common import TINY_ASSETS
 
@@ -92,22 +92,22 @@ class TestCliParseCommand(unittest.TestCase):
         self.assertRaises(ValueError, _parse_command, "?show-regions 0")
 
 
-class TestCliClient(unittest.TestCase):
-    """Tests for `smawg.cli._Client`.
+class TestCliHumanClient(unittest.TestCase):
+    """Tests for `smawg.cli._HumanClient`.
 
     That class is private, but having tests helps with fearless refactoring.
     """
 
     @patch("sys.stdout", new=StringIO())
     def test_exit_codes(self) -> None:
-        """`_Client.run()` should return 0 on 'quit' and 1 on EOF."""
+        """`_HumanClient.run()` should return 0 on 'quit' and 1 on EOF."""
         with patch("sys.stdin", new=StringIO("quit\n")):
             game = Game(TINY_ASSETS)
-            client = _Client(game)
+            client = _HumanClient(game)
             self.assertEqual(client.run(), 0)
         with patch("sys.stdin", new=StringIO("")):
             game = Game(TINY_ASSETS)
-            client = _Client(game)
+            client = _HumanClient(game)
             self.assertEqual(client.run(), 1)